Remove deposit slips
Working on the UI, I realized that we could make the system much more
understandable if it was a little closer to a prediction market. That
is, instead of the deposit slips, you deposit tokens into the vault and
receive your 'unit contract' (i.e., one that redeems into the thing if
the proposal passes and one if it fails).

This also makes the implementation simpler, reducing attack surface.